Move linked_list from json to detail
