+++ libcaf_core/caf/chunked_string.hpp
@@ -4,9 +4,8 @@
 
 #pragma once
 
-#include "caf/detail/build_config.hpp"
 #include "caf/detail/core_export.hpp"
-#include "caf/detail/json.hpp"
+#include "caf/detail/forward_list.hpp"
 #include "caf/detail/print.hpp"
 #include "caf/fwd.hpp"
 
@@ -22,10 +21,9 @@ namespace caf {
 /// Represents a chunked string as a linked list of string views.
 class CAF_CORE_EXPORT chunked_string {
 public:
-  using node_type = detail::json::linked_list_node<std::string_view>;
+  using node_type = detail::forward_list_node<std::string_view>;
 
-  using const_iterator
-    = detail::json::linked_list_iterator<const std::string_view>;
+  using const_iterator = detail::forward_list_iterator<const std::string_view>;
 
   explicit chunked_string(const node_type* head) noexcept : head_(head) {
     // nop
@@ -80,7 +78,7 @@ CAF_CORE_EXPORT std::ostream& operator<<(std::ostream& out,
 /// Builds a chunked string by allocating each chunk on a monotonic buffer.
 class CAF_CORE_EXPORT chunked_string_builder {
 public:
-  using list_type = detail::json::linked_list<std::string_view>;
+  using list_type = detail::forward_list<std::string_view>;
 
   using resource_type = detail::monotonic_buffer_resource;
